Allow settings from environment or local-only file
This enables the following:

  * specifying settings specific to Jenkins from the job without
    having to generate temporary files.

  * protect default settings from being modified by mistake.

  * simpler updates using git pull without merge conflict on settings
    for users.

+++ start.sh
@@ -1,8 +1,18 @@
-#!/bin/bash -e
-SCRIPT_DIR="$(cd "$(dirname "$0")"; pwd)"
+#!/bin/sh
+set -e
+
+# Find root folder
+export SCRIPT_DIR="$(cd "$(dirname "$0")"; pwd)"
 cd "${SCRIPT_DIR}"
 
-. raw-config.sh
+# Setup environment
+. ${SCRIPT_DIR}/settings.sh
+
+if [ -z "${RAW_VERSION}" ]
+then
+	echo "Please set RAW_VERSION, exiting..."
+	exit 1
+fi
 
 docker-compose -f docker-compose.yml rm -f postgres raw-storage raw-creds raw-executor raw-frontend
